WIP: Set the correct repo and add an API token
I'm not sure this is the correct type of token (it's an OAuth token from
my personal account).  The webhook setup seems to accpet a freeform
secret, but there isn't any documentation on the example in
serverless/components, so I don't know if I should be generating one to
include in the webhook setup...

If an OAuth token is what's required, then I'm not sure this one has the
correct permissions.  There aren't any permissions that pertain
specifically to issues in the config on Github.