WIP: Add most basic sinatra app for testing
Creates and extends a sinatra app that will be a webhook endpoints for a
GitHub webhook. The idea is to define an endpoint `/webhook` that takes
in issue data from Github when a new issue is created and adds a label
to issues that match a particular regex.

To make this work we need to define a regex, setup access to the github
API, parse the incoming json payload and add the label. This code does
that.