Make Puma run by default
We don't want to run sinatra using the default rack webserver but
instead want to use Puma. Puma is installed (check out the gemfile) but
we can use this little bit of config in the main sinatra file
`webhook.rb` to make sure Puma is used everytime you start the
application.

I've also added a little bit to the readme that reflects this way of
doing things.