Add a Dockerfile for the app
We had been deploying onto Heroku and so just needed a buildpack (ah,
the good old days!). Now we need a dockerfile so that we can build and
deploy a container to Kubernetes. This is my first go, based on some I
found on the internet. There are some tidy ups we can do, for instance:

* use multistage builds
* move some of the commands into a startup script
* use a non root user

but this is a reasonable start I think.