Make the app work on Kubernetes
There is a key change that needs to be made to make sure the app can run
on kubernetes which is in `webhook.rb` here. We need to make sure the
app binds to `0.0.0.0` not just to `localhost` which is [the
default](http://sinatrarb.com/configuration.html) hostname (see the
`:bind` section of that page).

I also added the docker command `EXPOSE` as that looks like best
practice.