Add full deployment instructions
I've updated the README with the full instructions to get the app
deployed in its current state. This involved making a couple of small
tidy up changes (delete `secrets.yaml` as we don't need it and get rid
of some comments in `webhook.rb`)

The things that I've left as is here but that might merit a bit of
change are:

* the "live" app deploys to the `-dev` namespace. The reason I have left
this as is is because I haven't had time to move the ECR repo. I'll
probably do that later but it is working now from `dev`.
* we are deploying the app manually, we could quite easily have a circle
CI job to do this, but is it worth it?